move the dialog to its own file and add some tests
